Move remaining deserialize logic to new Factory class
