+++ flint-spark-integration/src/main/scala/org/opensearch/flint/spark/FlintSparkIndex.scala
@@ -120,7 +120,7 @@ object FlintSparkIndex {
     builder.kind(index.kind)
     builder.options(index.options.optionsWithDefault.mapValues(_.asInstanceOf[AnyRef]).asJava)
 
-    // Optional index properties
+    // Index properties
     val envs = populateEnvToMetadata
     if (envs.nonEmpty) {
       builder.addProperty("env", envs.asJava)
