+++ flint-spark-integration/src/main/scala/org/opensearch/flint/spark/skipping/FlintSparkSkippingIndex.scala
@@ -80,7 +80,7 @@ class FlintSparkSkippingIndex(
     df.getOrElse(spark.read.table(tableName))
       .groupBy(input_file_name().as(FILE_PATH_COLUMN))
       .agg(namedAggFuncs.head, namedAggFuncs.tail: _*)
-      .withColumn(ID_COLUMN, sha1(col(FILE_PATH_COLUMN))) // TODO: no impact to just add it?
+      .withColumn(ID_COLUMN, sha1(col(FILE_PATH_COLUMN)))
   }
 }
